provide mvp on the main screen
